--- examples/datasets/.gitignore
@@ -1,2 +1,0 @@
-*.parquet
-*.ipc
