+++ py-polars/tests/slow/test_parquet.py
@@ -1,7 +1,9 @@
+import io
 import sys
 import typing
-from pathlib import Path
 
+import numpy as np
+import pandas as pd
 import pyarrow.dataset as ds
 import pytest
 
@@ -15,15 +17,7 @@ def test_struct_pyarrow_dataset_5796() -> None:
     num_rows = 2**17 + 1
 
     df = pl.from_records(
-        [
-            dict(  # noqa: C408
-                id=i,
-                nested=dict(  # noqa: C408
-                    a=i,
-                ),
-            )
-            for i in range(num_rows)
-        ]
+        [dict(id=i, nested=dict(a=i)) for i in range(num_rows)]  # noqa: C408
     )
 
     df.write_parquet("/tmp/out.parquet", use_pyarrow=True)
@@ -32,35 +26,21 @@ def test_struct_pyarrow_dataset_5796() -> None:
     assert_frame_equal(result, df)
 
 
-@pytest.mark.xfail(sys.platform == "win32", reason="Does not work on Windows")
-def test_sink_parquet_ipc(io_files_path: Path) -> None:
-    file = io_files_path / "small.parquet"
+def test_parquet_chunks_545() -> None:
+    cases = [1048576, 1048577]
 
-    dst = "/tmp/test_sink.parquet"
-    pl.scan_parquet(file).sink_parquet(dst)
-    with pl.StringCache():
-        assert pl.read_parquet(dst).frame_equal(pl.read_parquet(file))
+    for case in cases:
+        f = io.BytesIO()
+        # repeat until it has case instances
+        df = pd.DataFrame(
+            np.tile([1.0, pd.to_datetime("2010-10-10")], [case, 1]),
+            columns=["floats", "dates"],
+        )
 
-    dst = "/tmp/test_sink.ipc"
-    pl.scan_parquet(file).sink_ipc(dst)
-    with pl.StringCache():
-        assert pl.read_ipc(dst).frame_equal(pl.read_parquet(file))
+        # write as parquet
+        df.to_parquet(f)
+        f.seek(0)
 
-
-@pytest.mark.xfail(sys.platform == "win32", reason="Does not work on Windows")
-def test_fetch_union() -> None:
-    pl.DataFrame({"a": [0, 1, 2], "b": [1, 2, 3]}).write_parquet(
-        "/tmp/df_fetch_1.parquet"
-    )
-    pl.DataFrame({"a": [3, 4, 5], "b": [4, 5, 6]}).write_parquet(
-        "/tmp/df_fetch_2.parquet"
-    )
-
-    assert pl.scan_parquet("/tmp/df_fetch_1.parquet").fetch(1).to_dict(False) == {
-        "a": [0],
-        "b": [1],
-    }
-    assert pl.scan_parquet("/tmp/df_fetch_*.parquet").fetch(1).to_dict(False) == {
-        "a": [0, 3],
-        "b": [1, 4],
-    }
+        # read it with polars
+        polars_df = pl.read_parquet(f)
+        assert pl.DataFrame(df).frame_equal(polars_df)
